+++ config/model-watch.php
@@ -2,5 +2,7 @@
 
 // config for Mha/LaravelModelWatch
 return [
-
+    'collections' => [
+        'default' => \Mha\LaravelModelWatch\Collections\ExampleWatchUsers::class
+    ],
 ];