+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
 
 All notable changes to `laravel-model-watch` will be documented in this file.
+
+# Unreleased Changes
+
+ - Initial version of model:watch
+ - Monitor a single model
+ - Customise the interval and fields